/shore init (#43)
* Create the init function.

* Get a working init function.

* UX/UI fixes and a logical changes.

UX/UI:
1. Add a spinner when waiting for the `go` command to finish loading.
2. Add colored output for the `init` function.

Logical:
Change the dependency struct to depend on the `Project` instance.
The logic behind this change is there will always be one implementation of the `Project`. This is probably the only API that shore `core` requires. All other APIs are abstractions while this is the only concrenete implementation.

Another minor change is to make `params` default no matter what.

This means that running `shore render` with this file:
```
function(params)
  {}
```

Will work without modification.
The reasoning is that we would like to make `params` a defacto standard in shore packages.

* Fix renderer tests and add a default to params.

* Refactor project init to be testable.

* Always add `require.go` file.
Validate libraries are required before creating a new `import` statement.